Added INSTALL.txt (basic version at the moment)
Changes in people_image_downloader so that it traps the right error
when an image doesn't exist.

First attempt at the deployment. There has to be a better way though!
(I think there is ... this is just the beginning) NOTE: this only works
to deploy to my workstation at the moment.