Don't lookup the name on the page again, just reuse the person object
I can't really tell why this was looking the person up again. Maybe as
another check? It was causing a problem when parsing different names
so the downloader was saving the image for the wrong person.

It was downloading De-Anne Kelly and saving it for Craig Kelly. This
fixes that.