Better handling of a casting vote
This change is specifically to handle the situation encountered in
https://www.openaustralia.org.au/debates/?id=2018-09-17.26.2 - an
amendment was moved; the vote on the amendment tied, so the speaker
voted with the noes, which meant the amendment failed and the original
motion therefore passed.

As both the negative and the affirmative decisions happened with only
a single division, both the "agreed to" and "was negatived" phrases
are present. In order to catch this, we need to look to see where the
casting vote went: in this case, with the noes.

As well as being a more clear way to tell the outcome of the division,
this ensures that we are going to record the Speaker's vote correctly.